@@ -28,18 +28,19 @@
     "build:prod": "cross-env NODE_ENV=production yarn _build",
     "build:node": "yarn _babel && yarn build:ts",
     "build:ts": "tsc -p ./config/tsconfig.json",
+    "build:test": "tsc -p ./test/tsconfig.json",
     "build:browser": "webpack -c config/webpack.config.browser.js",
     "build:docs": "cross-env NODE_ENV=docs yarn _babel",
     "clean": "rm -rf lib/ dist/ coverage/ .nyc_output/ jsdoc/",
     "docs": "yarn build:docs && jsdoc -c ./config/.jsdoc.json --verbose",
-    "test": "yarn test:node && yarn test:integration && yarn test:browser",
+    "test": "yarn build:test && yarn test:node && yarn test:integration && yarn test:browser",
     "test:node": "yarn _nyc mocha --recursive 'test/unit/**/*.js'",
     "test:integration": "yarn _nyc mocha --recursive 'test/integration/**/*.js'",
     "test:browser": "karma start config/karma.conf.js",
     "fmt": "yarn eslint -c .eslintrc.js src/ --fix && yarn _prettier",
     "preversion": "yarn clean && yarn fmt && yarn build:prod && yarn test",
     "prepare": "yarn build:prod",
-    "_build": "yarn build:node && yarn build:browser",
+    "_build": "yarn build:node && yarn build:test && yarn build:browser",
     "_babel": "babel --extensions '.ts' --out-dir lib/ src/",
     "_nyc": "nyc --nycrc-path config/.nycrc",
     "_prettier": "prettier --ignore-path config/.prettierignore --write './**/*.js'"
@@ -76,23 +77,26 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.22.15",
-    "@babel/core": "^7.22.15",
+    "@babel/core": "^7.22.19",
     "@babel/eslint-parser": "^7.22.15",
     "@babel/eslint-plugin": "^7.22.10",
     "@babel/preset-env": "^7.22.15",
     "@babel/preset-typescript": "^7.22.15",
     "@babel/register": "^7.22.15",
-    "@definitelytyped/dtslint": "^0.0.177",
+    "@definitelytyped/dtslint": "^0.0.178",
     "@istanbuljs/nyc-config-babel": "3.0.0",
     "@stellar/tsconfig": "^1.0.2",
+    "@types/chai": "^4.3.6",
     "@types/detect-node": "^2.0.0",
     "@types/eventsource": "^1.1.2",
     "@types/lodash": "^4.14.198",
-    "@types/node": "^20.5.9",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.6.0",
     "@types/randombytes": "^2.0.0",
+    "@types/sinon": "^10.0.16",
     "@types/urijs": "^1.19.20",
-    "@typescript-eslint/parser": "^6.6.0",
-    "axios-mock-adapter": "^1.21.5",
+    "@typescript-eslint/parser": "^6.7.0",
+    "axios-mock-adapter": "^1.22.0",
     "babel-loader": "^9.1.3",
     "babel-plugin-istanbul": "^6.1.1",
     "buffer": "^6.0.3",
@@ -100,7 +104,7 @@
     "chai-as-promised": "^7.1.1",
     "chai-http": "^4.3.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.48.0",
+    "eslint": "^8.49.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.28.1",
@@ -127,7 +131,7 @@
     "nyc": "^15.1.0",
     "prettier": "^3.0.3",
     "randombytes": "^2.1.0",
-    "sinon": "^15.2.0",
+    "sinon": "^16.0.0",
     "sinon-chai": "^3.7.0",
     "taffydb": "^2.7.3",
     "terser-webpack-plugin": "^5.3.9",
@@ -142,7 +146,7 @@
     "bignumber.js": "^9.1.2",
     "eventsource": "^2.0.2",
     "randombytes": "^2.1.0",
-    "stellar-base": "^10.0.0-beta.1",
+    "stellar-base": "10.0.0-beta.1",
     "toml": "^3.0.0",
     "urijs": "^1.19.1"
   }
