@@ -5,8 +5,9 @@
 export * from "./soroban_rpc";
 
 // stellar-sdk classes to expose
-export { Server as SorobanClient } from "./server";
+export { Server as SorobanServer } from "./server";
 export { default as AxiosClient } from "./axios";
 export * from "./transaction";
+export * from "./contract_spec";
 
 export default module.exports;
